backend/requirements.txt
# AI Todo Chatbot Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
openai-agents>=0.1.0
sqlmodel>=0.0.14
alembic>=1.13.0
pydantic[email]>=2.6.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
httpx>=0.27.0
bcrypt>=4.1.0
PyJWT>=2.8.0
email-validator>=2.1.0
